(PC-9433) ___ WIP: search: Reorganize Algolia-related to code to ease migration
This commit introduces a new `pcapi.core.search` package. Like
`pcapi.core.mails` it has backends, which makes testing (and, more
importantly, the future migration to App Search) easier.

Code has been moved:
  process_eligible_offers -> search.reindex_offer_ids
  batch_processing_offer_ids_in_error -> search.index_offers_in_queue()
  batch_indexing_offers_in_algolia_by_venue -> search.index_venues_in_queue()

We use different Redis data structures for the queues of the Algolia
backend and the AppSearch backend: with Algolia, we use lists for
offer ids and venue ids queue, and a hashmap for "indexed_offers"
(because... legacy); with App Search, we use sets for all queues.